fix `getWeight` circular import
