+++ src/utils/usefulFunctions.ts
@@ -1,7 +1,4 @@
-import type { ProfessionName } from './gw2-data';
-import { Classes, Defense } from './gw2-data';
-
-function firstUppercase(text: string | undefined | null): string {
+export function firstUppercase(text: string | undefined | null): string {
   if (typeof text === 'undefined' || text === null || text === '') return '';
 
   const toUpper = (str: string) => str.charAt(0).toUpperCase() + str.slice(1).toLowerCase();
@@ -44,18 +41,6 @@ export const parsePriority = (text: number | string | null | undefined) =>
   parseNumber(text, null, false);
 export const parseBoss = (text: number | string | null | undefined) =>
   parseNumber(text, null, false);
-
-export const getWeight = (profession: ProfessionName) => {
-  // Calculate weight class
-  const { defense } = Classes[firstUppercase(profession) as ProfessionName];
-  if (defense === Defense.HEAVY) {
-    return 'Heavy';
-  }
-  if (defense === Defense.MEDIUM) {
-    return 'Medium';
-  }
-  return 'Light';
-};
 
 export const objectEntries = Object.entries as <Type extends object>(
   value: Type,
